requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
Werkzeug==2.3.7
scikit-learn==1.3.0
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
mysql-connector-python==8.1.0
